Implement comments counter
